Card Dealer Setup. BPI Card. Struct Card. Data table for cards. Card Names placeholder just so struct works. Card setup in construction script.
Card dealer actor which will contain all functionality for the setup of cards, shuffling of cards, and dealing of cards. These are all functions. Also the cards will be object pooled since there will be at least 60 probably so can just spawn them all in at the start of the game and then use them from this one card dealer actor which stores them all in an array called All Cards. This actor also has 2 other card arrays called player cards (the cards held by players both NPC and PC) and also Pile Cards which are the cards that are in the pile (both at the start of the game and during). There are also functions for removing and adding cards to/from the pile but these are not at all setup.

Setup cards function: main game name passed. Switch based on this. if i doubt it or rat screw then played cards are merely the all cards array since these games have players have all the cards in a deck. If it is action the player cards are only cards with a numerical value between 2 and 10 inclusive. We use the get card info function here from the card interface to check the numerical value. If it is not of this numerical range it is added to the pile cards array.

Shuffle cards function checks if pile cards is empty before shuffling them if not and then shuffles the player cards after both true for empty and shuffling pile cards.

Deal cards passes game name and players (array of actors) and then switches on the game name. Then this array will add the cards to each player as appropriate. This is not yet implemented but will be.

New card interface called BPI_Card which currently has one function to get the card info with return type struct_card.
This is obviously implemented in the card actor.

Struct_Card contains the following info: card name (enumerator of type card_names), static mesh, and numerical value (ace is 1 etc, king is 11, queen 12, jack 13, special cards will follow on from this). This numerical value is to enable card checking easier than checking if the card name is equal to something.

Card names eumerator currently just has test and test2 as names will need to fill this with all the names of the cards.

Cards data table. Empty as of now. Will contain all card info as data table since we know what all the cards are going to be before runtime so can do it this way.

Card set static mesh in construction script using the card info struct as a variable in the card actor which is exposed and instance editable.

The card dealer spawns all cards in on beginplay by getting all data table row names from the cards data table and then looping through these to get the data table row from the cards data table then spawning a card actor at 0,0,0 passing the out row as the card info and then setting the actor hidden in game and collision disabled. Then the current card in loop is added to the all cards array. When using cards we will then just have to move the cards to the correct location and rotation and then set them not hidden in game and collision enabled.