Changed BPI_Turn_Taker to BPI_Playing_Character so more general in its naming and function. Card dealer add card to pile and remove card from pile functions. Simplified some card dealer functions to only doing one thing in them. More functions in Playing Character interface.
The turn taker interface has now been changed to playing character interface and is found in the shared mechanics folder. This is so we are creating lots of interfaces when we can have just a few with the majority of the functions within it. This interface is now used for the removing and adding of cards to player hands.

Split the add card to pile and remove card to pile into multiple smaller functions since I was going to make them also add and remove from player hands but better to have those as separate functions in the card dealer since this will allow for easy card control and flow.

New functions then for adding cards to player hand and removing them from a character's hand. Pass the character to affect and the card to add / remove. These functions basically just call the corresponding function in the passed character's interface but this card dealer should act as an intermediary for card flow and control in the games.