New enum for all game names. Using game mode to store the current playing game. Card stacking. Some debug stuff in card pile functions. Actually finding ace of spades to add on first turn of I doubt it (this can be improved further)
Created a new enum that holds all potential games including mini games which will be used in the game mode allow easy access for what the current game being played is.

Game mode reference setup in player and NPCs on begin play.
They use this game mode reference in the take turn function to coordinate what to do for each game.

Stopped rat screw start game on begin play since I needed to test I doubt it, please Ben just use debug keys to setup and start games ty.

Setup some debug print strings in the card pile since was creating functionality for stacking cards.

Filled out the rest of the data table with the corresponding static meshes however two of the cards have the same static mesh since I think we created an extra special card after jess had already made the card meshes so might just need to get rid of one of them.

Created a testing material for the cards so can tell they are different which is just a base colour that gets randomised using a dynamic material instance in the card construction script.

When adding a card to the pile they are now stacked on top of each other. This is basic for now and yea but it works.

New function in the I Doubt It game manager called end player turn which does what was being done in the player and npc actors on take turn but instead consolidated in this function so that can change this and it will change it for all characters instead of having to change each individually.

Fixed the function for finding the ace of spades so it actually finds the ace of spades correctly and the character who has it.