Game mode changing mapping context. Leave current game function in game mode. Card hand visuals. Debug add card to hand for IDI
Game mode function set current playing game now also sets the player's input mapping context to be the corresponding game mapping context and removes the bar mapping context.
Debug function called leave current game which will remove the current game mapping context and add back the bar mapping context. This is a debug one for now since the player should not be able to leave a game whenever they want.

Two spheres on player card hand transform start and end and 2 arrows on these just so know direction.
Modified add card to hand in player and remove card from hand in player to add visuals for the card hands.
First we set the card not hidden in game. Then we looop through the cards in the our hand and set the actor location and rotation to be the following:
LOCATION:
Card hand transform start right vector inverted multiplied by ((card hand transform start world location distance to card hand transform end world location) / (length of hand - 1) ) * array index of hand. This is then added to the world location of the card hand transform start and is the location of the current card in the loop. The rotation is the camera world rotation, hand to change the actual rotation of the card actor in blueprint and so had to update rotation of the card pile in the world, may have to change the card pile stacking to be more modular as a result as well.
Then we attach this card to the camera keeping world for everything.
The removing card from hand event does this as well and it will work since the card is already removed before starting the loop through the hand.

Debug key 3 adds a card to player hand from the first NPC that has the tag I_Doubt_It_Participant and removes the card from their hand as well.