Card shaking. Card hovering and selecting update. Sorted input actions. New input action end turn. Modular max amount of cards selection.
New function in card interface for card shaking which is used to show when a selected card is hovered and is also used when selecting more cards than can be selected.
Card shaking uses relative rotation which is currently always 0 but if this changes for whatever reason then this will need to be changed in the shake card event. Also uses timeline obviously to lerp the rotation.

Organised the input actions.
Created a new input action for ending a turn.
This has replaced auto completion of turn for players in the take turn function.
This doesn't really do anything rn but is setup to use the different game names using the game mode. Also new bool in player which is if it this player's turn which allows them to end their turn. So can still do other actions like hovering and selecting but can't finish their turn until it is their turn.

Using a game mode to get current game name to set modular max amount of cards to select in select card function in currenthand component so can easily change it since different games have different max amounts between auction and I doubt it.

Fixed issue with cards not going to full selection height which is fixed now by setting the relative location before selecting to be the relative location before hovering plus the hover amount so it will always go to the full select amount.