I doubt it doubt call mechanic. Card pile remove card function setup.
3 new functions in i doubt it game manager which are associated with the doubting mechanic: Doubt check, doubt correct and doubt unsuccessful.
Just need to finish the doubt mechanic with tracking what card value should be played each turn.
New variables in the game manager: most recent played cards as an array of cards, and the current card value to play. This current value isnt fully setup yet.

New input action for calling doubt setup in player and in the i doubt it mapping context.
This input action calls the doubt function in the i doubt it game manager.

Remove card function in card pile which removes it from the card pile array. New parameter in function for removing card from pile in card dealer which is the pile to remove from which then calls this new function in the passed card pile.