Hold buttons for next and previous card selection. Advanced card hand placement method using spline now. Selecting and hovering with mouse movement and mouse press. New function in current hand component to check if passed card is a selected card/
Hold button next and previous card inputs so can now hold down the buttons to cycle through cards. Using same format as from Reign of the Abyss with the spell cycling. Using gates. All in player actor blueprint.
Input action for locking camera which is in the i doubt it mapping context. This uses a multigate to then set ignore look input true or false and show mouse cursor true or false. If we are locking then we start the custom event mouse to card event looping every 0.01 seconds. This timer is cleared when unlocking the camera.
The move to card function uses convert mouse location to world space to get the world location and direction and then using a line trace for objects of type card (had to create a new card object type and also created some new presets for card normal and card held which is changed in adding card to hand and removing card from hand. Also card actor enable collision is set true when adding card to hand and disabled again when removing from hand. This is done in the player since the player will have a hand of cards in front of them, could also do this for opponents as well in future.
For trace if we get a hit then we do a validated check for a new variable called card cursor hovered and if it is valid then we check if it is the same as the hit card so we only do the next things on cards that weren't the previously hovered card. If it is not the same then we call the is card selected function in the current hand component (new function) and if it isnt then we call the un hover card whilst held in the hit card and if so we carry on. The following is done as well if the hovered card is not valid from previous hovered we cast to BP_Card using the hit card and set the card cursor hovered to this hit actor and then call the hover check function in current hand component (new function).
If the trace doesnt hit anything we check if the card cursor hovered variable is valid and if so we call is card selected function in current hand component and if it is selected we just clear the card cursor hovered variable and if it isnt selected then we call the un hover card whilst held function in card and then clear the card cursor hovered variable.

Hover check sets the passed card to hovered card and then checks if it is selected to then shake it or hover it depending on if it is in the selected array.
Is card selected function in current hand component which just checks if passed card is contained within selected cards array.