Card sorting. Card sort methods.
New enum for card sorting methods currently type and value.

New function in current hand component for sorting the current hand by the passed card sort method.

New input action for sorting the current hand of cards. Using multigate to sort between suit and type.

Using custom sort method for suit type which just uses different local arrays of cards to store and then append on completion of loop. Only one loop required.

Using standard bubble sort for the numerical sort which is using the numerical value variable in the card info to sort the array into ascending order. There is a bubble sort for ascending order for numerical value and one for descending order.

Will probably optimise the bubble sort to be a more efficient sort algorithm since bubble is O(n2) and the cards in hand could get to like 30 or so.