Doubting advancement. Card Pile Advancement. Can only end turn once in the player's turn.
Can only doubt once now per turn at least in the player will need to make sure this is correct in the NPC when developing that.

New clear pile cards function in the card deal which does as says and calls the clear card pile function in the passed card pile.
When doubting successfully or unsuccessfully the pile cards are now cleared and the i doubt it card pile is cleared as well before adding to the specified player's hand.
Card pile now properly updates with doubting successful and unsuccessful.