Scroll wheel card selection
New input action for scroll wheel input. This action is added to the I doubt it input mapping context and allows the player to scroll through the cards in their hand. Tried to make it restricted so you couldn't scroll through the entire hand with a very high speed but couldn't arrive at a solution that worked so leaving it in for now since it doesn't break just means the player doesn't always see the animations if they go really fast through the hand.