Card Pile stack amount and random rotation deviation. Pile cards fixes and rat screw pile fixes. Card dealer fixes.
Card pile new variables: stack amount and random rotation deviation.
Stack amount is the space between the cards when they are stacked.
The random rotation deviation is the amount to randomly rotate each card in the card pile. This is done by dividing it by 2 and then getting a random float between the negative version of the half and the positive version of the half and then add that to the z rotation of the pile itself to set the card's rotation.

Pile card fixes via card dealer fixes:
Add card to pile array now only if card is not already in the pile cards array.
Rat screw manager now stores the card to play in the play card function as a local variable since it will get changed after calling the remove card from pile function so this was causing issues with card transfers between piles.
Then this local variable is used for the add card to pile function in this play card function in the screw game manager.