Remove card from pile optimisation. Doubt check only does through if pile is not empty. Doubt check. End turn check for player.
For remove card from visual pile function we now store the index of the card that is being removed. Then before we loop we check if this index is the same as the length to see if the card was removed from the top and if it wasn't then we alter the location. Also during the loop we now check if the card in the loop is above the card that was removed in its order in the array.

Doubt check now only goes through if the pile of cards that are used for I doubt it isnt empty.

Player end turn for I doubt it only goes through now if the selected cards isn't empty so the player cannot end their turn with no cards to play.