Level organisation. IDI playing positions functionality.
Organised the current test level to be in set folders.

Using target points for the playing positions for I doubt it.

Setup functionality in a macro to set these positions on begin play.

New function for moving the characters to positions. Loops through the array of playing characters and moves them to the corresponding position based on their order in the array, may have to change this in the future for specific positions based on the character themselves.