More npc decision making. NPC specific interface. Current card value to be played fixed. Play first turn function in turn manager fixed for IDI. Next and previous card fixes. IDI easy difficulty choice decision making done for now.
New interface specifically for NPCs with a function to get the specific difficulty the NPC has for the passed game.

IDI current card value to be played now starts at 1 instead of 0 since it was making it so that it would still be aces to be played after playing ace of spades.

Changed the IDI part of the start first turn function in the turn manager to use the end player turn in the IDI game manager which properly updates the card value to be played etc.

Changed some values in the player for the next and previous card button holding functionality.

Lots and lots and lots of work done for the IDI NPC decision making.
Done the easy difficulty decision making for now.
It is as follows:
Use the setup for getting all of the held unique cards, held correct cards, and held incorrect cards. Then switch on difficulty to then go to the easy difficulty.
Then calculate how many cards to play which is either the amount of cards that the player has left if they have less than 4 cards left or it is a randomly selected amount which is weighted using a random float in range between 0-1 and then doing bool checks for its value which can be changed to weight the different card amounts.
After this we calculate if we are going to bluff this round which is a simple random bool with weight.
Then we do a for loop with a break of 1 to the amount of cards to play. During the loop we first check if we are going to bluff then we check if we have any cards of the correct value if we do then we start looping through the cards of the correct type and find the first one that isn't already in the cards to play array, which we then add to the cards to play array. Now before adding it we do another check to see if it is already in the array since we could have gotten to the end of the array and so it could be in the array and if it is then we dont add it and instead check if we have reached the number of cards to play. If we have then we break the loop. If not we go to the next index. This ensures that we only play the number of cards to play or the number of cards that are of correct value if that is smaller than the amount of cards to play.

If we dont have any cards of the correct type then we must bluff so we do what we would normally do when bluffing with the amount of cards to be played staying the same.

Now onto if we are bluffing it goes as follows:
Check if we are bluffing -> we are so then check if both the unique cards and the cards of incorrect type are BOTH empty and if they are then we instead attempt to add a card of the correct type to the array as we would if we are not bluffing using the same functionality. If either are not empty then we proceed with a random bool with weight currently on 0.75 for playing an incorrect card and 0.25 for playing a unique card.
However before adding any cards we must check if the array we are going to add from (unique or incorrect) is empty respectively and if they are then we simple add the opposite type of card.
We use the same functionality of looping through the specified array (unique or incorrect) and check if they are within the cards to play array and if not we add them to the array but if we are we simply don't this time. This is the same functionality for both unique and incorrect cards.

The following are card playing rules that ALL ENEMY DIFFICULTIES WILL FOLLOW:
If we have only one card left then play that card.
If the length of the correct cards is 4 then play all 4 of those cards.
Those are currently the only rules that override all of the enemy difficulties.
There are more comments in the IDI decide function in the NPC Decision mind component that explain my thought process for the different difficulty strategies.