3 new variables in NPC decision mind component IDI Main decide
Amount of card weightings which are float arrays which are used for the calculating of how many cards should be played in a hand. There is one for each difficulty.

We now only get player cards of the correct type for the medium and hard difficulties.

We only get the opponent cards of the correct type as well now for the medium and hard difficulties which is stored in a map correlating the card to the opponent difficulty so that we can use these dependent on how observant the enemy holding it is.

Will work more on the medium and hard difficulties tomorrow or the day after.