Expert difficulty NPC decision making for IDI.
Some more variables: Expert N. Cards Held by Experts which is a counter including players holding the correct cards.
Filled the expert difficulty card amount weightings with some floats for now.
Expert initial bluffing mistake randomisation - chance for a mistake where we bluff even though it might not be the correct play.
Expert extra card unique chance - the chance for doing a unique card over a incorrect card when adding a card that isn't the correct type to the cards to play.

Changed the easy difficulty adding incorrect or unique cards checks to use the same way expert difficulty does as well but need to properly test all this eventually since some of it will need testing with cards running out.

Bluffing for expert NPCs is as follows: if the NPC has no cards of correct type or if the expert initial bluffing mistake randomisation is true.
If we are not bluffing then we first check if both the opponents don't have any cards of the correct type and the player doesn't have any cards of the correct type. If both of these are empty then we add a card of the correct type in order of the for loop playing and check if it is valid so if it is valid then we check if the card is contained in the cards to play array and if not then add it to the array. If the card is not valid then we do a check for if we should do unique cards of incorrect cards using the expert extra card unique chance and a random float value. Depending on this we then check if the corresponding array is empty for unique or incorrect and then loop through the corresponding array attempting to add the card to the array if it is not contained in the cards to play array. We break the array if we get to the end of the corresponding array then we try to add a card of the opposite array.

If we are bluffing then we check if the opponent card correct type -> difficulty array contains any cards held by expert NPCs and if so then we start looping through the cards and check if it is held by an expert NPC and if so then we increment the counter for cards held by expert NPCs or players. On completion of the array or if the array has no cards held by expert NPCs then we loop through the player's hand and increment the counter.
On completion of this loop we then calculate the amount of cards to play by doing 4 - the counter for correct cards held by experts and player.
Then we start looping through with 1 to the amount of cards to play. First in the loop we check if the unique and incorrect cards are both empty and if sot then we check if we have any cards of the correct type (this check should never produce true due to previous checks and if it were true then this means we have no cards which means we wouldn't play any cards anyway). If it is not empty which it shouldn't be then we add the first card from the cards of the correct type array to the cards to play array and then stop the loop.
The other branch starts with doing a check for if we should play a unique card or an incorrect card using the expert extra card unique chance variable to determine if we should play a unique card.
Then we do the loops for adding either a unique card or incorrect card to the cards to play array with checks for the last index of the opposite array being in the cards to play array which would mean we need to add from the opposite array. If both arrays already have all their cards in the cards to play array then we just break the loop and those are then the cards that we will play.