IDI Medium difficulty.
Medium difficulty is very similar to the expert difficulty with some more mistakes and randomisation as well as fewer knowledge of other player's cards.
Some specific medium difficulty local variables:
Medium initial bluffing mistake float which is currently at 25% for the chance to bluff even if we weren't going to though this might go down quite a bit after testing.
Medium extra card unique randomisation float which is the chance to play a unique card over a incorrect card.
Number of cards held by non-experts integer which stores how many cards of the correct type are held by non-experts and not the player.

The differences between medium and expert NPC decision making for I doubt it include:
Higher chance for bluffing to occur even if we dont need to bluff and have correct cards to play.
This NPC only knows the correct cards of non-expert NPCs and not the player so they don't know completely who has what cards so will make riskier plays without having all of the data to back this play up and so it will fail more than the expert NPC will when bluffing.

May change this more and will also need to thoroughly test all of the different difficulties for the NPCs as well soon.