First turn fixes for IDI and Rat Screw. Debugging NPC decision.
Had a order of execution wrong in the first turn function in the turn manager for setting the character turn index which was being set after calling the end player turn function when it should be called before it.
Also fixed it for rat screw as well since it was the same issue I think.

Did some debugging for the NPC decision for IDI using some local variables for determining if the print strings of the different difficulties should print or not.
Added a lot of print strings will continue debugging later and hopefully get it fully working as I want.