I Doubt It Game Over Check adjustment. Adjusted some expert difficulty values and made them play a little riskier. Doubt check additional check. IDI Side choice setup.
Check game over has another prerequisite for the player to win: they must have played a round with 0 cards to play.
This means that if they are doubted on their last cards played and have to pick up then they will not win yet which is how the game winning should function.

Removed doubt check function in IDI game manager and instead added a check in the doubt check function to see if any one has doubted this round already and if they have then we don't do the rest of the doubt check. This will make it easier for NPCs to doubt and removes an unnecessary function in the game manager and function call in the player and NPCs.

Setup some local variables in the NPC decision mind component doubt option function for speed, awareness and accuracy for each different difficulty.