Doubt Decide NPC functionality.
Local Variables for Reaction speed, accuracy, awareness, frequency and no info.
There are also local variables for the number of correct cards held by this NPC and the number of correct cards in the pile excluding those played in the previous round.

Reaction speed checks are done at the end of branches and are the final check between doubting and not doubting which is supposed to serve as the NPC being fast enough to doubt before the next player plays their cards.

Accuracy is how well the NPC can track how many cards of the correct type for the previous round are in the card pile.
Awareness is how often the NPC will take note of their own cards in relation to the cards that have just been played.
Frequency is the first check of how often the NPC will actually even attempt to doubt however even if this check fails there are some other checks that will happen that would make the NPC doubt anyway including if the NPC has 4 cards of the correct type which then leads to an awareness check which if passed then leads to a reaction speed test.
No info variables are used when frequency passes, we have no correct cards in hand and there are no correct cards in the pile. There are only no info checks for medium and easy difficulty with expert difficulty just proceeding to not doubt at that point.

It's all commented and decently complicated however there are only 4 main checks it just depends on their order and the information determining sometimes if it is a guaranteed doubt depending on the difficulty as well.

Also made this function private so it cannot be accessed by outside actors and so we only use the optional action decide function passing the correct variables for the game and difficulty. Will need to go through various actors and change their function access specifiers to have more data encapsulation and control over the data flow in the game as well as just generally making the game more professional programming-wise.