Doubt functionality for NPCs actually fully setup now. Needs testing though and balancing.
New variable in the IDI game manager called NPC Doubt Timers which is an array of timer handles. This is added to after calling the Doubt Check function with a timer in the NPC decision mind component.

New function in the IDI game manager: Clear NPC Doubt Timers. This loops through the array and clears and invalidates each timer handle and then clears the array on completion of this loop.
This function is called on doubt check passing the initial 2 checks and is also called at the start of the end player turn function.

At the end of the end player turn function we now loop through all of the playing characters and then check if they are an NPC if they are then we get their difficulty and call the optional action decide function in the NPC mind decision component which then allows each of the NPCs to possibly doubt this round of cards that has just played.

At the of the optional action decide function in the NPC decision mind component after calling the doubt decide function we then check if we are going to doubt this round and if so then we set a timer of Doubt Check in the IDI game manager of a timer randomly between 0.25 and 0.45 so it feels a little more organic and so there is a chance that they are still too slow and that the next player will play their cards. After setting this timer we then add it to the NPC doubt timer array in the IDI game manager.

All of this should work need to properly test it but wanted to commit it now just in case I don't get around to testing it tonight.