Altered a lot of the variables for the doubt decide. Fixed NPCs doubting using new function in NPC opponent interface. Debugging doubt decide function.
Nerfed all of the reaction speeds and frequencies for the different difficulties.

Fixed logic issue with doubt check being called with a timer.
It is fixed by creating a new function in the opponent interface for doubting which is called with a timer instead of calling the doubt check directly in the game manager since need to pass the NPC as a parameter so this new function in the interface does that.

So now the timers work properly now as well.

Some print strings setup for debugging the doubt decide function with a new bool local variable in the function for if we want to debug or not.

Need to keep debugging and testing to see if everything is working properly.