Bug Fixing + Slap AI
Fixed player not being able to play multiple cards in a row
Fixed turns changing rapidly when a round is won
Fixed NPC going twice for no apparent reason

Commented over the play card and deal shared cards function

Implemented NPC slapping when a slap is available. First does a random weight with bool to see if a slap will be attempted and then slaps in an amount of time determined by difficulty. Want to vary the time is taken to slap a little bit later on.