Unique Card Function within IDI Game Manager. New enum for unique card names. Removed beast from card names.
New function in the IDI game manager which is called when adding a card of unique suit to the pile.

This function switches on the passed unique card name enum and then does functionality accordingly.
Multiple local variables used in this function as well.

New enum for just the unique card names.
Removed beast card from all enums.

In end player turn function check if the card being added to the pile is of a unique suit and if so then call the unique card function passing the unique card name using a select based on the enum card names.

Need to make it so that there is priority for the unique cards so playing multiple unique cards only makes one of the card's effects active.