Doubt Check considers Joker now. Doubt Check considers Angel and Devil now as well. Doubted round bool only resets on next player turn now. Some unique card fixes. New functions in turn manager for getting character index previous and next including direction. Fixing doubt giving to correct player. New variable in I Doubt It manager for most recent player.
The doubt check function in the I Doubt It manager now also checks for jokers in the cards. It also checks now for angels and devils active now as well which adjust how the doubt correct / incorrect works.

The doubted round bool in the IDI game manager is now only reset on the next player turn instead of also being reset in another function.

Fixed some of the unique card function character index references.

New functions in the turn manager for getting the next character index and one for getting the previous character index which also use the direction bool in the turn manager.

Fixed giving doubt to the round player for doubt correct by using a new variable in for the most recent played player which is updated on end player turn function using the passed player variable in this function. This is used in various different functions now in the IDI game manager which fixes some issues.