Fixed compile issue with card info. NPC Doubt frequency dependent on number of players as well as rounds gone without doubting. More debugging messages for doubting.
Fixed a compile error for card info struct in the card dealer by resetting the variable in the card actor and then refreshing the nodes in the card dealer. Hope this isn't a recurring issue since its not hard to fix just annoying.

The doubting frequency for each difficulty is now also divided by the number of playing characters - 1 so that they don't all have really high doubt frequencies but can still if fewer characters playing. Additionally the frequency is increased by adding the current round - the doubted round index all divided by 100 so the more cards in the pile / the longer gone without successfully doubting the higher the frequency of their doubting increases.