Current turn UI functionality setup in turn manager. Using placeholder arrow for current turn indicator. Doubt call widget. Pause, resume and cancel player turn functions in i doubt it game manager. Changed how turns are managed for I doubt it specifically. Fixed a lot of unique card action issues for I Doubt It.
New widget called current turn indicator.
Functionality for setup and moving of current turn indicator. This is done in turn manager actor.
Using placeholder arrow asset from https://uxwing.com/arrow-down-icon/ for now for the current turn indicator.

Doubt call widget which is currently just text that says: DOUBT!!!
Functionality for this widget in the I Doubt It game manager for both setup and moving.

Changed the take turn function in the NPC to use a new function for I Doubt It specifically which is now what is on a timer. This timer is still stored in the IDI game manager and is managed in various functions for pausing, clearing and resuming functions in the IDI manager.

End player turn doesn't automatically call check if game over function and will only call it if the cards played don't contain a unique card.
Therefore have more control over turn management since calling the check if game over function in the actual unique card action function.

There was quite a few issues with the unique card action function that have now been fixed however the famine bug still isn't fixed yet and needs to be fully implemented. Pretty sure the rest of them are fixed however.
Unique card actions that surround player turns works with current turn indicator.