Current turn indicator bounce animation functionality. Doubt calling for NPCs not done for most recent player.
Current turn indicator has a bounce animation using a timeline which is activated when it reached its location and rotation for the current turn.

When looping through playing characters for the NPCs to call the optional side choice function it also checks if it is the most recent player and only calls if not so that the NPC who just played cannot call doubt on their own cards played. Think this may have fixed the issue I found with doubting but need to do more testing and debugging for overall timing etc.