Current round card value indicator / widget. Played cards round indicator / widget. Card Counter indicator / widget.
Widget for current card value of the round which is displayed above the pile card and rotated to face the player's camera. This shows the players what card value should be played for the round.

Played cards for the round indicator which is a widget that will eventually be a speech bubble thing to show what cards the player has just supposedly played.

Card counter indicator widget which is a placeholder icon and number of how many cards the player has.
Functionality for all of these indicators are in the I Doubt It game manager. Using a map for player to created card counter widget and then updating its value in the end player turn function after removing the corresponding cards from the played player's hand.