Auction round end function fixes. Current hand fixes for selected cards for auction. Drawing and playing characters active for round change for auction.
Fixed quite a few mistakes in the round end function in the auction game manager concerning the scoring of players and whether or not it was a draw for the round.
Also in the end round function we also now call the clear selected cards function in each player's current hand component after we have already calculated their scores for the round so that they are playing the correct cards each round (more so for the player).

New way of checking if it is a draw. After calculating the highest score we then loop through the player scores again and check if any of them match the highest score. If they do we add them to a local array called drawing indexes. After this we now check if the length of this array is greater than 1 meaning that more than 1 player had the highest score for the round so there is a draw.
Then we loop through the players and check if they are a drawing player using the drawing indexes array and if they aren't then we set their corresponding boolean in the playing characters map to be false so that they won't play a card for the repeating round. Then after this loop we call NPC select all cards for round and then check if the player is a drawing player and if so we call take turn with a 1 second delay and if not we call end round with a 2 seconds delay.

In NPC select all cards for round function we check whilst looping that the player should play a card this round using their corresponding boolean in the playing characters map.

This commit in summary is drawing for auction and some other fixes and changes surrounding auction.